feat: Move Cache tests to their own project
