+++ platform/src/abstractions/tests/Platform.Cache.Tests/Cache/WhenRedisDistributedCacheGetsString.cs
@@ -2,7 +2,7 @@
 using StackExchange.Redis;
 using Xunit;
 using Xunit.Abstractions;
-namespace Platform.Tests.Cache;
+namespace Platform.Cache.Tests.Cache;
 
 public class WhenRedisDistributedCacheGetsString(ITestOutputHelper testOutputHelper) : RedisDistributedCacheTestBase(testOutputHelper)
 {
