+++ platform/src/abstractions/tests/Platform.Cache.Tests/Cache/WhenRedisDistributedCacheSetsObjects.cs
@@ -3,7 +3,7 @@ using StackExchange.Redis;
 using Xunit;
 using Xunit.Abstractions;
 // ReSharper disable NotAccessedPositionalProperty.Global
-namespace Platform.Tests.Cache;
+namespace Platform.Cache.Tests.Cache;
 
 public class WhenRedisDistributedCacheSetsObjects(ITestOutputHelper testOutputHelper) : RedisDistributedCacheTestBase(testOutputHelper)
 {
@@ -18,8 +18,8 @@ public class WhenRedisDistributedCacheSetsObjects(ITestOutputHelper testOutputHe
     {
         var actualValues = Array.Empty<KeyValuePair<RedisKey, RedisValue>>();
         Database
-            .Setup(d => d.StringSetAsync(It.IsAny<KeyValuePair<RedisKey, RedisValue>[]>(), When.Always, CommandFlags.None))
-            .Callback((KeyValuePair<RedisKey, RedisValue>[] v, When _, CommandFlags _) =>
+            .Setup(d => d.StringSetAsync(It.IsAny<KeyValuePair<RedisKey, RedisValue>[]>(), StackExchange.Redis.When.Always, CommandFlags.None))
+            .Callback((KeyValuePair<RedisKey, RedisValue>[] v, StackExchange.Redis.When _, CommandFlags _) =>
             {
                 actualValues = v;
             })
@@ -46,7 +46,7 @@ public class WhenRedisDistributedCacheSetsObjects(ITestOutputHelper testOutputHe
         const string key = nameof(key);
         const string value = nameof(value);
         Database
-            .Setup(d => d.StringSetAsync(It.IsAny<KeyValuePair<RedisKey, RedisValue>[]>(), When.Always, CommandFlags.None))
+            .Setup(d => d.StringSetAsync(It.IsAny<KeyValuePair<RedisKey, RedisValue>[]>(), StackExchange.Redis.When.Always, CommandFlags.None))
             .ThrowsAsync(new RedisConnectionException(ConnectionFailureType.UnableToConnect, "Unable to connect to Redis"))
             .Verifiable(Times.Once);
