+++ platform/src/abstractions/tests/Platform.Cache.Tests/Cache/WhenRedisDistributedCacheSetsString.cs
@@ -2,7 +2,7 @@
 using StackExchange.Redis;
 using Xunit;
 using Xunit.Abstractions;
-namespace Platform.Tests.Cache;
+namespace Platform.Cache.Tests.Cache;
 
 public class WhenRedisDistributedCacheSetsString(ITestOutputHelper testOutputHelper) : RedisDistributedCacheTestBase(testOutputHelper)
 {
@@ -11,7 +11,7 @@ public class WhenRedisDistributedCacheSetsString(ITestOutputHelper testOutputHel
     public async Task ShouldSetValueInCache(string key, string value)
     {
         Database
-            .Setup(d => d.StringSetAsync(key, value, null, false, When.Always, CommandFlags.None))
+            .Setup(d => d.StringSetAsync(key, value, null, false, StackExchange.Redis.When.Always, CommandFlags.None))
             .ReturnsAsync(true)
             .Verifiable(Times.Once);
 
@@ -27,7 +27,7 @@ public class WhenRedisDistributedCacheSetsString(ITestOutputHelper testOutputHel
         const string key = nameof(key);
         const string value = nameof(value);
         Database
-            .Setup(d => d.StringSetAsync(key, value, null, false, When.Always, CommandFlags.None))
+            .Setup(d => d.StringSetAsync(key, value, null, false, StackExchange.Redis.When.Always, CommandFlags.None))
             .ThrowsAsync(new RedisConnectionException(ConnectionFailureType.UnableToConnect, "Unable to connect to Redis"))
             .Verifiable(Times.Once);
